Fixed access marco error and modified demog output var from cell array to table

+++ matlab/programs/updateIDList.m
@@ -22,6 +22,15 @@ function updateIDList( varargin, spss_flag )
 
 %%%%%%%%-------BEGIN CODE-------------%%%%%%%%%%%%%%
 
+%Update spss demog file
+if nargin < 1
+    varargin = ' ';
+    spss_flag=0;
+elseif nargin < 2
+    spss_flag=0;
+end
+
+
 
 %Added this so we could overwrite if needed
 if(strcmpi('force',varargin))
@@ -30,24 +39,19 @@ else
     run_flag = false;
 end
 
-%Update spss demog file
-if nargin < 2
-    spss_flag=0;
-end
-
 
 data_dir = [pathroot 'tmp\'];
 
 orig=pwd;
 %fpath = [pathroot 'db/master id list.xlsx'];
 %fpath = 'L:/Summary Notes/Data/matlab/db/master id list.xlsx';
-fpath = 'c:\kod\Neuropsych_preproc\matlab\db\splashDemo2.xlsx';
+fpath = 'c:\kod\Neuropsych_preproc\matlab\db\ALL_SUBJECTS_DEMO.xlsx';
 
 if ~exist(fpath,'file')
     error('No master list found, see help'); % Just in case
 end
 
-hardcopy = 'Y:\Front End (For Copying)\Jon\Protect 2 - Front End (Jon).accdb';
+hardcopy = 'Y:\Front End (For Copying)\Protect DB Front End.accdb';
 %Check original file's date
 org_file=dir(strcat(fpath));
 org_file=org_file.date;
@@ -58,17 +62,20 @@ db_file=db_file.date;
 
 %Compare dates buffer added to negate redundancy
 if datenum(org_file)+.03<=datenum(db_file) || run_flag
-    disp('Database is out of date! Don''t worry we can put a bird on it....');
+    disp('Database is out of date! Don''t worry we can put a bird on it....'); %Databases are over!
     
     %Connect to database and export mast list
     h= actxserver('Access.Application');
     invoke(h,'OpenCurrentDatabase',hardcopy); %Protect DB
-    invoke(h.DoCmd,'RunMacro','ExportDemographics'); %Macro is currently only executable on my (Jon's) computer
+    invoke(h.DoCmd,'RunMacro','exportDemog_noRunMacro'); %Macro is currently only executable on my (Jon's) computer
     h.Visible = 0;
     
+    %Garbage collection
+    h.Quit;
+    delete(h);
     %Copy it to the L drive
     %cd('L:/Summary Notes/Data/matlab/db/')
-    copyfile(fpath,'L:/Summary Notes/Data/matlab/db/splashDemo2.xlsx');
+    copyfile(fpath,'L:/Summary Notes/Old L-drive/Data/matlab/db/ALL_SUBJECTS_DEMO.xlsx');
 %     movefile('master id list.xlsx','master id list_old_bckup.xlsx'); %Create backup
 %     movefile('master_id_list.xlsx','master id list.xlsx');
     disp('All DONE, press any key to continue!');
@@ -81,7 +88,7 @@ end
 %directory used for the older Jan code
 
 %Load data into matlab
-filename = [pathroot 'db/splashDemo2.xlsx'];
+filename = [pathroot 'db/ALL_SUBJECTS_DEMO.xlsx'];
 [~,~,data] = xlsread(filename);
 
 %Strip header information
@@ -101,6 +108,22 @@ end
 id_nums=[data{:,index.ID}]';
 tmp = zeros(length(id_nums),1);
 
+
+%Determine what study(ies) each subject participated in
+studies = {'SUICIDE' 'SUICID2' 'PROTECT' 'PROTECT2' 'AFSP'  'LEARN' ...
+    'EXPLORE' 'B_SOCIAL'};
+
+for i = 1:length(studies)
+    study_index = index.(studies{i});
+    enrolled_in_study=~cellfun(@sum,cellfun(@isnan, data(:,study_index),'UniformOutput',false)); %Make logical vector
+    index_update = length(fieldnames(index)) +1;
+    index.([studies{i} '_BIN']) = index_update; %Add binary header for final data struct
+    data(:,index_update) = num2cell(enrolled_in_study);
+
+end
+    
+
+
 %Find where there are duplicates
 for i = 2:length(id_nums)
     if id_nums(i)==id_nums(i-1)
@@ -109,40 +132,58 @@ for i = 2:length(id_nums)
 end
 idx=find(tmp);
 count = 0;
-for i=1:length(idx)
-    %Not so clean workaround look into making this more robust
-    idx(i)=idx(i)-count;
-    
-    %PROTO field
-    data{idx(i)-1,index.PROTO}=strcat(data{idx(i)-1,index.PROTO},'/',data{idx(i),index.PROTO});
-    
-    %Combine dates adjust for SPSS UTF-16 error
-    if ~strcmpi(data{idx(i)-1,index.Consent},data{idx(i),index.Consent}) && strcmpi(data{idx(i)-1,index.Consent},'NaN')
-        data{idx(i)-1,index.Consent}=strcat(data{idx(i)-1,index.Consent},'--',data{idx(i),index.Consent});
-    end
-    
-    if ~strcmpi(data{idx(i)-1,index.DateofTermination},data{idx(i),index.DateofTermination}) && strcmpi(data{idx(i)-1,index.DateofTermination},'NaN')
-        data{idx(i)-1,index.DateofTermination}=strcat(data{idx(i)-1,index.DateofTermination},'--',data{idx(i),index.DateofTermination});
-    end
-    
-    %Delete the duplicate row
-    data(idx(i),:)=[];
-    
-    count=count+1;
-end
 
+%Only run if duplicates are found
+if ~isempty(idx)
+    
+    %Find the duplicates -- if any though there should not be!
+    [C,ia,ic] = unique(id_nums);
+    diff_ia = diff(ia);
+    dup_idx = find(diff(ia)>1);
+    reoccurance=diff_ia(find(diff(ia)>1));
+    
+    if ~isempty(dup_idx) %This is a massive pain as working with cells is no bueno, but if need be you can figure it out
+        error('There is a duplicate! You need to fix this or talk to Josh again!')
+        return
+        for i = 1:length(dup_idx)
+            tmp = cell(1,length(data(dup_idx(i),:)));
+            for j = 1:reoccurance(i)
+                ct = j-1;
+                cell_idx = cellfun(@isnan, data(dup_idx(i)+ct,:), 'UniformOutput', false);
+                B = ~any(cellfun(@isnan,A(:,4:end)),2);
+                tmp(cell_idx) = data(dup_idx(i)+ct,cell_idx);
+            end
+        end
+    end
+end
 %Clean up "other paitent" data, rmv 'NaN' entries
-other_idx=strcmp(data(:,index.PATTYPE),'OTHER PATIENT');
+other_idx=strcmpi(data(:,index.PATTYPE),'OTHER PATIENT');
 data(other_idx,:)={'NaN'};
 qnan = cellfun(@any,cellfun(@isnan,data,'UniformOutput',0));
 qnan = ( qnan |  strcmp('NaN',data) ); % get rid of 'NaN' entries
 data(qnan) = {''}; % replace NaN's with zeros
 emptyCells = cellfun('isempty', data);
-data(all(emptyCells,2),:) = [];
+
+%5/13/2016 talk to Josh, this was removing everyone that had an empty cell,
+%people had empty cells becasue they had Nan's in various columns, (dates,
+%ect)
+%data(all(emptyCells,2),:) = [];
+data(emptyCells(:,1),:) = [];
+
+%Convert data into table format
+final_headers = fieldnames(index)';
+data=cell2table(data,'VariableNames',final_headers);
+
+%Final clean up of data, some numbers are still strings, remove any nans
+%that appear
+data=convert_string_to_double(data,{'MAXLETHALITY'});
+data=convert_string_to_double(data,{'GROUP12467'});
 
 % Archive and save it
 cd(data_dir)
-copyfile('demogs_data.mat', 'demogs_data_archive.mat');
+if exist('demogs_data.mat','file')
+    copyfile('demogs_data.mat', 'demogs_data_archive.mat');
+end
 save([data_dir 'demogs_data'],'data');
 
 
@@ -191,3 +232,7 @@ fclose all;
 cd(orig);
 end
 
+function T=convert_string_to_double(T,my_field)
+T.(my_field{:}) = str2double(T.(my_field{:}));
+T.(my_field{:})(isnan(T.(my_field{:}))) = 0;
+end
